Multi_Agent_Application/requirements.txt
streamlit>=1.22.0
pandas
requests
snowflake-snowpark-python